Use meta_box_cb option for Issues taxonomy
Replaces the standard multi-selecting post editing metabox for Issues
with the single-selecting version. The metabox is then altered again by
PeriodicalPress_Admin::add_remove_metaboxes(), but now if that went
wrong for any reason the correct metabox would still appear, just in
the wrong place and with the wrong title.