Created Singleton parent class
The new parent class is used for all Singleton classes in the plugin,
reducing duplicated code and making the individual class’s code quicker
to navigate.

Issue: this commit uses late static binding, which means PHP 5.2 isn’t
supported by the plugin. But it wasn’t prior to this anyway.